+++ backend/deepchecks_monitoring/bgtasks/model_version_topic_delete.py
@@ -1,16 +1,13 @@
 import threading
-from ssl import SSLContext
 from typing import Optional
 
 import pendulum as pdl
-from aiokafka.admin import AIOKafkaAdminClient
-from aiokafka.admin import __version__ as aiokafka_version
-from aiokafka.client import AIOKafkaClient
 from kafka.errors import KafkaError, UnknownTopicOrPartitionError
 from sqlalchemy import delete, select
 from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from backend.deepchecks_monitoring.utils.other import ExtandedAIOKafkaAdminClient
 from deepchecks_monitoring.logic.keys import get_data_topic_name
 from deepchecks_monitoring.monitoring_utils import TimeUnit
 from deepchecks_monitoring.public_models import Organization
@@ -24,39 +21,6 @@ __all__ = ['ModelVersionTopicDeletionWorker', 'insert_model_version_topic_delete
 
 QUEUE_NAME = 'model version topic delete'
 DELAY = TimeUnit.HOUR.value * 3
-
-
-class ExtandedAIOKafkaAdminClient(AIOKafkaAdminClient):  # pylint: disable=missing-class-docstring
-    # pylint: disable=super-init-not-called
-    def __init__(self, *, loop=None,
-                 bootstrap_servers: str = 'localhost',
-                 client_id: str = 'aiokafka-' + aiokafka_version,
-                 request_timeout_ms: int = 40000,
-                 connections_max_idle_ms: int = 540000,
-                 retry_backoff_ms: int = 100,
-                 metadata_max_age_ms: int = 300000,
-                 security_protocol: str = 'PLAINTEXT',
-                 ssl_context: Optional[SSLContext] = None,
-                 api_version: str = 'auto',
-                 sasl_mechanism='PLAIN',
-                 sasl_plain_username=None,
-                 sasl_plain_password=None):
-        self._closed = False
-        self._started = False
-        self._version_info = {}
-        self._request_timeout_ms = request_timeout_ms
-        self._client = AIOKafkaClient(
-            loop=loop, bootstrap_servers=bootstrap_servers,
-            client_id=client_id, metadata_max_age_ms=metadata_max_age_ms,
-            request_timeout_ms=request_timeout_ms,
-            retry_backoff_ms=retry_backoff_ms,
-            api_version=api_version,
-            ssl_context=ssl_context,
-            security_protocol=security_protocol,
-            connections_max_idle_ms=connections_max_idle_ms,
-            sasl_mechanism=sasl_mechanism,
-            sasl_plain_username=sasl_plain_username,
-            sasl_plain_password=sasl_plain_password)
 
 
 class ModelVersionTopicDeletionWorker(BackgroundWorker):
